Implement class-based views with access control and permission handling
- Created class-based views for course management (Create, Update, Delete, and List views)
- Introduced mixins (OwnerMixin, OwnerEditMixin, OwnerCourseMixin) to handle common logic for filtering and form validation by course owner
- Added Django's LoginRequiredMixin and PermissionRequiredMixin to restrict access based on user authentication and permissions
- Defined permissions for course views to ensure that only authorized users can add, edit, delete, or view courses
- Configured URL patterns for course management actions and integrated these with main project URLs
- Created templates (list.html, form.html, delete.html) to support the course views for listing, creating, editing, and deleting courses, providing user-friendly UIs
